Added various heuristics.
The general rule of thumb is if we think the data maybe misplaced,
regardless what the mapping quality states, then we should preserve
qualities.  The intention here is that realigning this sample to an
updated reference will likely need the qualities for any misplaced
reads.

Our heuristics are partial (see the long TODO message), but so far
cover variability in insertion length, excessive depth, excessive
amounts of soft-clipping and high portion of low mapping quality
(perhaps this just comes under excessive depth?).