Slight improvement to the indel size detection.
We now report back the reference min/max for STR rather than
pos+STR_length.  This differs when the STR is an insertion, which was
causing more qualities to be included in the output than were strictly
necessary.

Also tightened the indel size multiply for -9 so it's more distinct
from -7.  As before, it can be overridden using e.g. -9 -i1.1,2