Changed return values for -h / no-args.  Fixes #5
These both now return exit code zero.

We consider just "crumble" to be a request to see the usage and a
non-error, but "crumble -Q 50" for example to be an incomplete command
line and thus erroneous.

Also hide the usage reporting on error.  It typically scrolls the
useful error message off the screen, so this is now just a trivial one
liner.