Added sequence error correction (-a -A) options.
This is possibly over-egging the pudding a bit, but it is disabled by
default.  Use this option can significantly reduce the size of the
sequence data (due to more predictability and fewer differences).
With -A it also further reduces the size of the quality data.

Overall these can reduce file sizes by a further ~20%.