Add better support for PacBio HiFi (CCS).
There is a -Y FLOAT option to control the fraction of reads containing
indels before we attempt to do STR analysis on any read within that
pileup.  This helps reduce excess time spent in the STR caller for
technologies that have higher indel error rates, except where we
believe indels are likely genuine.  -Y0.1 approximately doubles the
speed of Crumble on HiFi data.

The -U INT option specifies a maximum cap on quality value.  This
permits us to limit the range of qualities used for the reads that we
wish to keep, avoiding the excess of 93 distinct quality values.  This
definitely reduces file size, but it's debatable how valid it is as
some tools allegedly have custom code for recognising Q93.  Code
exists to retain this, but commented out for now as it appears to be
slightly detrimental to variant calling accuracy and poorer on size
too.

Added -y STRING option for specifying instrument.  Current strings
recognised are "illumina" which does nothing as the defaults were
tuned for that, and "pbccs" which adjusts -u, -U and -Y.