Huge speed up to mask_LC_regions.
On long read data (tested with a 20kb PB-CCS library) this can get
called over 1000 times per read.  We don't cache the results, which we
could do, but for simplicity it is now limited to only computing the
min/max mask for a maximum of 500bp either side of the specified
position.

This is an order of magnitude quicker.