Add a -k/-K options to keep (preserve) specific quality values.
This can be useful for keeping Q93 (for example) on PacBio data, which
is utilised by some calling pipelines.

The -k option requests preserving qualities when we deem them to be
obvious homozygous but a high value is present in a discrepant call.
For example this could be a somatic mutation.

The -K option is a more explicit keep-qual where we always store the
qualities regardless of whether there are bases that differ.  It's not
expected this mode would be used commonly.

Finally the -N option upgrades any single base to keep to an entire
column of bases to keep.  The logic for this is if we have quantised
down a whole set of quals with -l10 -c30 -u50 so we get Q10s and Q50s
only present for the data agreeing with our hypothesised call, but
still have 1 or 2 non-call bases at Q93, then it can adversely skew
the call confidence.  In that case we'd rather keep all the original
confidence values too.